use consistent approach to validate arguments
