+++ packages/services/src/adapters/storage/index.ts
@@ -1,0 +1,1 @@
+export * from "./storage.adapter.js";
